Integrated JavaAnalyzer enhancements plus began plug-in refactoring.
There were many enhancements done to the JavaAnalyzer in the past 40 or
so commits merged in from another repo focusing on improving the Java
features and capabilities.  Some of these broke the DB analyzer.
Updated the DB Analyzer to work with these changes and also incorporated
some of UI fixes.