more plug-in refactoring
added a plugin rest controller (for giving plug-in data to the
front-end) and a converter controller to do the SVG-to-PNG conversion.
moved the controllers for the db and java to their respective packages.
added a plugins module that defines the PluginInterface implemented by a
plug-in class in each of the plug-in modules (java and db)
still have a bit more work to do, but definitely an improvement