More refactoring!
Cleaned up which packages some classes live in.
Additional work on the model visitor.
Moved graph creation outside of the model. Now lives in the graphBuilder
package. This implements a Factory pattern to select the correct builder
based on the type of graph.
Extracted the upstream/downstream dependency depth calculations out of
the model and into the DependencyDepthVisitor. This gets called by the
graph builders for updating the exclude lists when generating graphs
with reference depths.